requirements.txt
Flask>=2.0
flask-cors>=3.0
requests>=2.28
python-owasp-zap-v2.4>=1.0
requests[socks]>=2.28
